The submitting and Addingg tasks
I wrote an event when the form was sumbitted
I wrote a function that make sure that the client's text isnt empty or spaces - if it return true I can continue with get the whole data from the form - the text , the priority(select) and the date time.
All the data getting in an object called task (line 40) and it gets an specific Id with the const variable of Ids.
I created a function that work when I need to add a task that bulid a containar div that hold 3 dives for the priority,date and text and put the data from the object task into them .
after this it displays the task on the DOM.